Add LaunchSettingsValueProviderBase
1. Add a `LaunchSettingsValueProviderBase` that serves as a public base class for getting and setting property values in the debug launch settings. We will make use of this internally, and it will also be of use to the ASP.NET team as they create property pages that work with Codespaces.
2. Rework `ActiveLaunchProfileCommonValueProvider` and `ActiveLaunchProfileExtensionValueProvider` to make use of the new type.
3. Make the base types public.
