+++ src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/Properties/InterceptedProjectProperties/DebugPropertyPage/LaunchSettingsValueProviderBase.cs
@@ -1,0 +1,65 @@
+﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
+
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.VisualStudio.ProjectSystem.Debug;
+
+namespace Microsoft.VisualStudio.ProjectSystem.Properties
+{
+    public abstract class LaunchSettingsValueProviderBase : InterceptingPropertyValueProviderBase
+    {
+        private readonly UnconfiguredProject _project;
+        private readonly ILaunchSettingsProvider _launchSettingsProvider;
+        private readonly IProjectThreadingService _projectThreadingService;
+
+        public LaunchSettingsValueProviderBase(UnconfiguredProject project, ILaunchSettingsProvider launchSettingsProvider, IProjectThreadingService projectThreadingService)
+        {
+            _project = project;
+            _launchSettingsProvider = launchSettingsProvider;
+            _projectThreadingService = projectThreadingService;
+        }
+
+        public override Task<string> OnGetEvaluatedPropertyValueAsync(string propertyName, string evaluatedPropertyValue, IProjectProperties defaultProperties)
+        {
+            return GetPropertyValueAsync(propertyName);
+        }
+
+        public override Task<string> OnGetUnevaluatedPropertyValueAsync(string propertyName, string unevaluatedPropertyValue, IProjectProperties defaultProperties)
+        {
+            return GetPropertyValueAsync(propertyName);
+        }
+
+        public override Task<string?> OnSetPropertyValueAsync(string propertyName, string unevaluatedPropertyValue, IProjectProperties defaultProperties, IReadOnlyDictionary<string, string>? dimensionalConditions = null)
+        {
+            _projectThreadingService.RunAndForget(async () =>
+            {
+                ILaunchSettings launchSettings = await _launchSettingsProvider.WaitForFirstSnapshot(Timeout.Infinite);
+
+                var writableLaunchSettings = launchSettings.ToWritableLaunchSettings();
+                if (SetPropertyValue(propertyName, unevaluatedPropertyValue, writableLaunchSettings))
+                {
+                    await _launchSettingsProvider.UpdateAndSaveSettingsAsync(writableLaunchSettings.ToLaunchSettings());
+                }
+            },
+options: ForkOptions.HideLocks,
+unconfiguredProject: _project);
+
+            // We've intercepted the "set" operation and redirected it to the launch settings.
+            // Return "null" to indicate that the value should _not_ be set in the project file
+            // as well.
+            return Task.FromResult<string?>(null);
+        }
+
+        private async Task<string> GetPropertyValueAsync(string propertyName)
+        {
+            ILaunchSettings launchSettings = await _launchSettingsProvider.WaitForFirstSnapshot(Timeout.Infinite);
+
+            return GetPropertyValue(propertyName, launchSettings) ?? string.Empty;
+        }
+
+        public abstract string? GetPropertyValue(string propertyName, ILaunchSettings launchSettings);
+
+        public abstract bool SetPropertyValue(string propertyName, string value, IWritableLaunchSettings launchSettings);
+    }
+}
